[refactor] setting 패키지 코딩 컨벤션 적용
