+++ gotbetter/src/main/java/pcrc/gotbetter/detail_plan/data_access/entity/DetailPlan.java
@@ -15,7 +15,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import pcrc.gotbetter.participant.data_access.entity.ParticipantInfo;
-import pcrc.gotbetter.setting.BaseTimeEntity;
+import pcrc.gotbetter.setting.common.BaseTimeEntity;
 
 @Entity
 @Table(name = "DetailPlan")
