+++ gotbetter/src/main/java/pcrc/gotbetter/plan/data_access/entity/Plan.java
@@ -17,7 +17,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import pcrc.gotbetter.participant.data_access.entity.ParticipantInfo;
-import pcrc.gotbetter.setting.BaseTimeEntity;
+import pcrc.gotbetter.setting.common.BaseTimeEntity;
 
 @Entity
 @Table(name = "Plan")
