+++ gotbetter/src/main/java/pcrc/gotbetter/plan_evaluation/data_access/entity/PlanEvaluation.java
@@ -10,7 +10,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import pcrc.gotbetter.setting.BaseTimeEntity;
+import pcrc.gotbetter.setting.common.BaseTimeEntity;
 
 @Entity
 @Table(name = "PlanEvaluation")
