+++ gotbetter/src/main/java/pcrc/gotbetter/setting/http_api/ApiErrorView.java
@@ -1,57 +1,59 @@
 package pcrc.gotbetter.setting.http_api;
 
-import lombok.Getter;
-import lombok.ToString;
-import org.springframework.util.ObjectUtils;
-
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import org.springframework.util.ObjectUtils;
+
+import lombok.Getter;
+import lombok.ToString;
+
 @Getter
 @ToString
 public class ApiErrorView {
-    private final List<Error> errors;
+	private final List<Error> errors;
 
-    public ApiErrorView(List<MessageType> messageTypes) {
-        this.errors = messageTypes.stream().map(Error::errorWithMessageType).collect(Collectors.toList());
-    }
+	public ApiErrorView(List<MessageType> messageTypes) {
+		this.errors = messageTypes.stream().map(Error::errorWithMessageType).collect(Collectors.toList());
+	}
 
-    public ApiErrorView(GotBetterException exception) {
-        this.errors = Collections.singletonList(Error.errorWithException(exception));
-    }
+	public ApiErrorView(GotBetterException exception) {
+		this.errors = Collections.singletonList(Error.errorWithException(exception));
+	}
 
-    public ApiErrorView(MessageType messageType, String message) {
-        this.errors = Collections.singletonList(Error.errorWithMessageTypeAndMessage(messageType, message));
-    }
+	public ApiErrorView(MessageType messageType, String message) {
+		this.errors = Collections.singletonList(Error.errorWithMessageTypeAndMessage(messageType, message));
+	}
 
-    @Getter
-    @ToString
-    public static class Error {
-        private final String errorType;
-        private final String errorMessage;
+	@Getter
+	@ToString
+	public static class Error {
+		private final String errorType;
+		private final String errorMessage;
 
-        public static Error errorWithMessageType(MessageType messageType) {
-            return new Error(messageType.name(), messageType.getMessage());
-        }
+		public static Error errorWithMessageType(MessageType messageType) {
+			return new Error(messageType.name(), messageType.getMessage());
+		}
 
-        public static Error errorWithMessageTypeAndMessage(MessageType messageType, String message) {
-            return new Error(messageType.name(), message);
-        }
+		public static Error errorWithMessageTypeAndMessage(MessageType messageType, String message) {
+			return new Error(messageType.name(), message);
+		}
 
-        public static Error errorWithException(GotBetterException gotBetterException) {
-            return new Error(gotBetterException);
-        }
+		public static Error errorWithException(GotBetterException gotBetterException) {
+			return new Error(gotBetterException);
+		}
 
-        private Error(String errorType, String errorMessage) {
-            this.errorType = errorType;
-            this.errorMessage = errorMessage;
-        }
+		private Error(String errorType, String errorMessage) {
+			this.errorType = errorType;
+			this.errorMessage = errorMessage;
+		}
 
-        private Error(GotBetterException gotBetterException) {
-            this.errorType = ObjectUtils.isEmpty(gotBetterException.getType()) ? gotBetterException.getStatus().getReasonPhrase() :
-                    gotBetterException.getType();
-            this.errorMessage = gotBetterException.getMessage();
-        }
-    }
+		private Error(GotBetterException gotBetterException) {
+			this.errorType =
+				ObjectUtils.isEmpty(gotBetterException.getType()) ? gotBetterException.getStatus().getReasonPhrase() :
+					gotBetterException.getType();
+			this.errorMessage = gotBetterException.getMessage();
+		}
+	}
 }
