+++ gotbetter/src/main/java/pcrc/gotbetter/setting/http_api/GotBetterException.java
@@ -1,16 +1,17 @@
 package pcrc.gotbetter.setting.http_api;
 
-import lombok.Getter;
 import org.springframework.http.HttpStatus;
+
+import lombok.Getter;
 
 @Getter
 public class GotBetterException extends RuntimeException {
-    private final HttpStatus status;
-    private final String type;
+	private final HttpStatus status;
+	private final String type;
 
-    public GotBetterException(MessageType messageType) {
-        super(messageType.getMessage());
-        this.status = messageType.getStatus(); //ex) 404
-        this.type = messageType.name(); //ex) NOT_FOUND
-    }
+	public GotBetterException(MessageType messageType) {
+		super(messageType.getMessage());
+		this.status = messageType.getStatus(); //ex) 404
+		this.type = messageType.name(); //ex) NOT_FOUND
+	}
 }
