+++ gotbetter/src/main/java/pcrc/gotbetter/user/data_access/entity/User.java
@@ -15,7 +15,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import pcrc.gotbetter.setting.BaseTimeEntity;
+import pcrc.gotbetter.setting.common.BaseTimeEntity;
 import pcrc.gotbetter.user.login_method.login_type.RoleType;
 
 @Entity
@@ -26,62 +26,62 @@ import pcrc.gotbetter.user.login_method.login_type.RoleType;
 @DynamicUpdate
 public class User extends BaseTimeEntity {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "user_id", nullable = false)
-    private Long userId;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "user_id", nullable = false)
+	private Long userId;
 
-    @Column(nullable = false)
-    private String username;
+	@Column(nullable = false)
+	private String username;
 
-    @Column(nullable = false)
-    private String email;
+	@Column(nullable = false)
+	private String email;
 
-    private String profile;
+	private String profile;
 
-    @Column(name = "role_type", length = 20, nullable = false)
-    @Enumerated(EnumType.STRING)
-    private RoleType roleType;
+	@Column(name = "role_type", length = 20, nullable = false)
+	@Enumerated(EnumType.STRING)
+	private RoleType roleType;
 
-    @Column(name = "refresh_token")
-    private String refreshToken;
+	@Column(name = "refresh_token")
+	private String refreshToken;
 
-    @Column(name = "fcm_token")
-    private String fcmToken;
+	@Column(name = "fcm_token")
+	private String fcmToken;
 
-    @Builder
-    public User(
-        Long userId,
-        String username,
-        String email,
-        String profile,
-        RoleType roleType,
-        String refreshToken,
-        String fcmToken
-    ) {
-        this.userId = userId;
-        this.username = username;
-        this.email = email;
-        this.profile = profile;
-        this.roleType = roleType;
-        this.refreshToken = refreshToken;
-        this.fcmToken = fcmToken;
-    }
+	@Builder
+	public User(
+		Long userId,
+		String username,
+		String email,
+		String profile,
+		RoleType roleType,
+		String refreshToken,
+		String fcmToken
+	) {
+		this.userId = userId;
+		this.username = username;
+		this.email = email;
+		this.profile = profile;
+		this.roleType = roleType;
+		this.refreshToken = refreshToken;
+		this.fcmToken = fcmToken;
+	}
 
-    public void updateRefreshToken(String refreshToken) {
-        this.refreshToken = refreshToken;
-    }
+	public void updateRefreshToken(String refreshToken) {
+		this.refreshToken = refreshToken;
+	}
 
-    public void updateUsername(String username) {
-        this.username = username;
-    }
+	public void updateUsername(String username) {
+		this.username = username;
+	}
 
-    public void updateFcmToken(String fcmToken) {
-        this.fcmToken = fcmToken;
-    }
+	public void updateFcmToken(String fcmToken) {
+		this.fcmToken = fcmToken;
+	}
 
-    public void updateById(String userId) {
-        this.updateCreatedById(userId);
-        this.updateUpdatedById(userId);
-    }
+	public void updateById(String userId) {
+		this.updateCreatedById(userId);
+		this.updateUpdatedById(userId);
+	}
 }
