+++ gotbetter/src/main/java/pcrc/gotbetter/user/data_access/entity/UserSet.java
@@ -11,7 +11,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import pcrc.gotbetter.setting.BaseTimeEntity;
+import pcrc.gotbetter.setting.common.BaseTimeEntity;
 
 @Entity
 @Table(name = "UserSet")
@@ -20,25 +20,25 @@ import pcrc.gotbetter.setting.BaseTimeEntity;
 @DynamicInsert
 @DynamicUpdate
 public class UserSet extends BaseTimeEntity {
-    @Id
-    @Column(name = "user_id", nullable = false)
-    private Long userId;
+	@Id
+	@Column(name = "user_id", nullable = false)
+	private Long userId;
 
-    @Column(name = "auth_id", nullable = false)
-    private String authId;
+	@Column(name = "auth_id", nullable = false)
+	private String authId;
 
-    @Column(nullable = false)
-    private String password;
+	@Column(nullable = false)
+	private String password;
 
-    @Builder
-    public UserSet(Long userId, String authId, String password) {
-        this.userId = userId;
-        this.authId = authId;
-        this.password = password;
-    }
+	@Builder
+	public UserSet(Long userId, String authId, String password) {
+		this.userId = userId;
+		this.authId = authId;
+		this.password = password;
+	}
 
-    public void updateById(String userId) {
-        this.updateCreatedById(userId);
-        this.updateUpdatedById(userId);
-    }
+	public void updateById(String userId) {
+		this.updateCreatedById(userId);
+		this.updateUpdatedById(userId);
+	}
 }
